refactor: split CLIcore_UI_execute.c into focused sub-modules
CLIcore_UI_execute.c was 3,934 lines and contained unrelated
responsibilities: debugging utilities, text-rewriting transforms,
I/O redirection handlers, and the main dispatch loop.

Split into four files each under 1,000 lines:

  CLIcore_UI_execute.c        (2,423 lines → ~1,000 with dispatch only)
  CLIcore_UI_execute_debug.c  (135 lines)
    - write_tracedebugfile()
    - CLI_execute_string()

  CLIcore_UI_execute_preproc.c (653 lines)
    - cli_check_unquoted_restricted_symbols()
    - cli_split_logical_op()
    - cli_rewrite_stream_pipe()
    - cli_split_pipe()
    - cli_split_semicolon()
    - cli_rewrite_dot_source()

  CLIcore_UI_execute_redir.c   (1,001 lines)
    - cli_run_external()
    - cli_handle_subshell()
    - cli_handle_herestring_late()
    - cli_handle_stderr_redir()
    - cli_handle_input_redir()
    - cli_handle_output_redir()
    - cli_handle_herestring_early()
    - cli_handle_background()

A new internal header CLIcore_UI_execute_internal.h provides
forward declarations for all helpers — replaces static linkage
with explicit prototypes so functions can live in separate TUs.

Verified: 276/276 CLI robustness tests pass.
