+++ src/cli/libmilkscript/CLIcore_UI_execute.c
@@ -1,15 +1,19 @@
 /**
  * @file CLIcore_UI_execute.c
  *
- * @brief Core CLI execution engine and dispatcher
+ * @brief Core CLI command dispatcher.
  *
- * Architecture Overview:
- * This file implements the central command loop execution engine.
- * The primary entry point is `CLI_execute_line()`, which is responsible for
- * taking a raw input string and running it through a sequence of script-like
- * preprocessing steps (variable expansion, arithmetic, aliases, control flow)
- * before dispatching it to either a native C module command or an external
- * system binary.
+ * Contains CLI_execute_line(), the central command-
+ * processing loop. Helper functions are in sub-modules:
+ *
+ *  CLIcore_UI_execute_debug.c
+ *      write_tracedebugfile, CLI_execute_string
+ *  CLIcore_UI_execute_preproc.c
+ *      Text transforms: logical ops, pipe splitting,
+ *      semicolons, dot-source rewrite
+ *  CLIcore_UI_execute_redir.c
+ *      I/O redirection, subshell, background,
+ *      external command runner
  */
 
 #include <stdio.h>
@@ -30,6 +34,7 @@
 #include "CLIcore/cli_calc_parser.h"
 #include "CLIcore_script.h"
 #include "CLIcore_UI_execute.h"
+#include "CLIcore_UI_execute_internal.h"
 
 #include <errno.h>
 #include <fnmatch.h>
@@ -62,1521 +67,6 @@
 
 extern void yy_scan_string(const char *);
 extern int  yylex_destroy(void);
-
-
-
-
-/**
- * @brief Dump the circular code-trace buffer to a log
- *        file for post-mortem debugging.
- *
- * Writes every non-empty entry from the dctestptarr[]
- * ring buffer into a timestamped log file named
- * milk-codetracepoint.<PID>.log. Each entry includes
- * the source file, line number, function name, message,
- * and the full function-call stack at that tracepoint.
- *
- * This is called on abnormal exit (signal handler or
- * explicit user request) so the developer can inspect
- * the execution path that led to a crash.
- */
-errno_t write_tracedebugfile()
-{
-    pid_t thisPID = getpid();
-
-    char fname[STRINGMAXLEN_FILENAME];
-    WRITE_FILENAME(fname, "milk-codetracepoint.%05d.log", thisPID);
-
-    printf("Writing output trace to file %s\n", fname);
-    printf("dctestptinit = %d\n", dctestptinit);
-
-    FILE *fp = fopen(fname, "w");
-    if(fp != NULL)
-    {
-        for(uint64_t i = 0; i < CODETESTPOINTARRAY_NBCNT; i++)
-        {
-            long j = (i + dctestptcnt) % CODETESTPOINTARRAY_NBCNT;
-
-            uint64_t index =
-                dctestptarr[j].loopcnt * CODETESTPOINTARRAY_NBCNT + j;
-
-            if(dctestptarr[j].line != 0)
-            {
-                char timestring[TIMESTRINGLEN];
-                mkUTtimestring_nanosec(timestring, dctestptarr[j].time);
-
-                // extract last word
-                char str[STRINGMAXLEN_FULLFILENAME];
-                strcpy(str, dctestptarr[j].file);
-                char *lastword = strrchr(str, '/') + 1;
-
-                fprintf(fp,
-                        "T %6ld %s %-20s %6d %-20s  %s\n",
-                        index,
-                        timestring,
-                        lastword,
-                        dctestptarr[j].line,
-                        dctestptarr[j].func,
-                        dctestptarr[j].msg);
-                fprintf(fp,
-                        "       FTRACE %d ",
-                        dctestptarr[j].funclevel);
-                for(int level = 0; level < dctestptarr[j].funclevel;
-                        level++)
-                {
-                    fprintf(fp,
-                            " (%d) >> %ld:%s",
-                            dctestptarr[j].linestack[level],
-                            dctestptarr[j].fcntstack[level],
-                            dctestptarr[j].funcstack[level]);
-                }
-                fprintf(fp, "\n\n");
-
-                //printf("%s\n", p + 1);
-            }
-        }
-        fclose(fp);
-    }
-
-    return RETURN_SUCCESS;
-}
-
-
-
-/**
- * @brief Execute a command string without clobbering
- *        the current command line.
- *
- * Saves the current data.CLIcmdline, replaces it with
- * @cmd, runs CLI_execute_line(), then restores the
- * original. This allows re-entrant command execution
- * (e.g. from within scripts, trap handlers, or
- * on_update callbacks) without losing the outer
- * command context.
- *
- * @param cmd  Null-terminated command string to execute
- * @return RETURN_SUCCESS on success, error code on
- *         failure
- */
-errno_t CLI_execute_string(const char *cmd)
-{
-    char save_cmdline[STRINGMAXLEN_CLICMDLINE];
-    strncpy(save_cmdline, data.CLIcmdline, STRINGMAXLEN_CLICMDLINE);
-    strncpy(data.CLIcmdline, cmd, STRINGMAXLEN_CLICMDLINE - 1);
-    data.CLIcmdline[STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-    errno_t ret = CLI_execute_line();
-    strncpy(data.CLIcmdline, save_cmdline, STRINGMAXLEN_CLICMDLINE);
-    return ret;
-}
-
-
-/**
- * @brief Detect shell meta-characters outside quotes.
- *
- * Scans @cmdline character by character, tracking
- * single-quote, double-quote, and backslash-escape
- * state. Returns 1 if any restricted symbol
- * (; < > | [ ] ( ) & * $) appears outside quotes.
- *
- * The '=' character is only allowed in valid
- * assignment context (after an identifier).
- *
- * This gate decides whether the line must be
- * dispatched to /bin/sh (restricted symbols present)
- * or can be handled by the milk parser directly.
- *
- * @param cmdline  Command line to scan
- * @return 1 if restricted symbols found, 0 if clean
- */
-static int cli_check_unquoted_restricted_symbols(
-    const char *cmdline
-)
-{
-    int in_squote = 0;
-    int in_dquote = 0;
-    int esc = 0;
-    
-    /* '[' and ']' removed to allow stream
-     * slicing syntax (e.g. im[0:19,10:29]). */
-    const char *restricted = ";<>|()*&$";
-    
-    int word_start = 1;
-    int valid_assign_prefix = 0; 
-    
-    for (int i = 0; cmdline[i] != '\0'; i++) {
-        char c = cmdline[i];
-        
-        if (esc) {
-            esc = 0;
-            word_start = 0;
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (c == '\\') {
-            esc = 1;
-            word_start = 0;
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (in_squote) {
-            if (c == '\'') in_squote = 0;
-            continue;
-        }
-        
-        if (in_dquote) {
-            if (c == '"') in_dquote = 0;
-            continue;
-        }
-        
-        if (c == '\'') {
-            in_squote = 1;
-            word_start = 0;
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (c == '"') {
-            in_dquote = 1;
-            word_start = 0;
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (isspace(c)) {
-            word_start = 1;
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (strchr(restricted, c) != NULL) {
-            return 1;
-        }
-        
-        if (c == '=') {
-            if (!valid_assign_prefix) {
-                return 1;
-            }
-            valid_assign_prefix = 0;
-            continue;
-        }
-        
-        if (word_start) {
-            if (isalpha(c) || c == '_'
-                || c == '@') {
-                valid_assign_prefix = 1;
-            } else {
-                valid_assign_prefix = 0;
-            }
-            word_start = 0;
-        } else {
-            if (!isalnum(c) && c != '_'
-                && c != '.' && c != '@') {
-                valid_assign_prefix = 0;
-            }
-        }
-    }
-    
-    return 0;
-}
-
-/**
- * @brief Split command line at top-level && or ||.
- *
- * Scans for the first unquoted, un-nested && or ||
- * operator. If found, executes the left side, then
- * conditionally executes the right side based on
- * success/failure.
- *
- * @param[out] retval  Set to the return value if
- *                     the line was consumed
- * @return 1 if the line was consumed, 0 otherwise
- */
-static int cli_split_logical_op(errno_t *retval)
-{
-    const char *src = data.CLIcmdline;
-    int depth = 0;
-    int in_sq = 0;
-    int in_dq = 0;
-    int split_pos = -1;
-    int op_len = 0;
-    int op_is_and = 0;
-
-    for (int si = 0; src[si] != '\0'; si++)
-    {
-        char c = src[si];
-        if (c == '\'' && !in_dq)
-        {
-            in_sq = !in_sq;
-        }
-        else if (c == '"' && !in_sq)
-        {
-            in_dq = !in_dq;
-        }
-        else if (!in_sq && !in_dq)
-        {
-            if (c == '(')
-            {
-                depth++;
-            }
-            else if (c == ')' && depth > 0)
-            {
-                depth--;
-            }
-            else if (depth == 0
-                     && c == '&'
-                     && src[si + 1] == '&')
-            {
-                split_pos = si;
-                op_len = 2;
-                op_is_and = 1;
-                break;
-            }
-            else if (depth == 0
-                     && c == '|'
-                     && src[si + 1] == '|')
-            {
-                split_pos = si;
-                op_len = 2;
-                op_is_and = 0;
-                break;
-            }
-        }
-    }
-    if (split_pos < 0)
-    {
-        return 0;
-    }
-
-    char right[STRINGMAXLEN_CLICMDLINE];
-    const char *rp = src + split_pos + op_len;
-    while (*rp == ' ' || *rp == '\t')
-    {
-        rp++;
-    }
-    strncpy(right, rp,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    right[STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-    char left[STRINGMAXLEN_CLICMDLINE];
-    strncpy(left, data.CLIcmdline,
-            (size_t) split_pos);
-    left[split_pos] = '\0';
-    strncpy(data.CLIcmdline, left,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    data.CLIcmdline[
-        STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-    errno_t lret = CLI_execute_line();
-    int ok = (cli_last_retval == 0);
-    int run_right =
-        (op_is_and && ok)
-        || (!op_is_and && !ok);
-    if (run_right)
-    {
-        strncpy(data.CLIcmdline, right,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        data.CLIcmdline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-        *retval = CLI_execute_line();
-        return 1;
-    }
-    *retval = lret;
-    return 1;
-}
-
-
-/**
- * @brief Rewrite stream pipeline |> syntax.
- *
- * Transforms "stream |> cmd args" into
- * "cmd stream args" and re-executes.
- *
- * @param[out] retval  Set to the return value if
- *                     the line was consumed
- * @return 1 if the line was consumed, 0 otherwise
- */
-static int cli_rewrite_stream_pipe(
-    errno_t *retval
-)
-{
-    const char *src = data.CLIcmdline;
-    int depth = 0;
-    int in_sq = 0;
-    int in_dq = 0;
-    int gpipe = -1;
-
-    for (int si = 0; src[si] != '\0'; si++)
-    {
-        char c = src[si];
-        if (c == '\'' && !in_dq)
-        {
-            in_sq = !in_sq;
-        }
-        else if (c == '"' && !in_sq)
-        {
-            in_dq = !in_dq;
-        }
-        else if (!in_sq && !in_dq)
-        {
-            if (c == '(')
-            {
-                depth++;
-            }
-            else if (c == ')' && depth > 0)
-            {
-                depth--;
-            }
-            else if (depth == 0
-                     && c == '|'
-                     && src[si + 1] == '>')
-            {
-                gpipe = si;
-                break;
-            }
-        }
-    }
-    if (gpipe < 0)
-    {
-        return 0;
-    }
-
-    char lhs[STRINGMAXLEN_CLICMDLINE];
-    strncpy(lhs, data.CLIcmdline,
-            (size_t) gpipe);
-    lhs[gpipe] = '\0';
-    {
-        int e = gpipe - 1;
-        while (e >= 0
-               && (lhs[e] == ' '
-                   || lhs[e] == '\t'))
-        {
-            lhs[e--] = '\0';
-        }
-    }
-    const char *rhs =
-        data.CLIcmdline + gpipe + 2;
-    while (*rhs == ' ' || *rhs == '\t')
-    {
-        rhs++;
-    }
-    const char *sp = strchr(rhs, ' ');
-    char newcmd[STRINGMAXLEN_CLICMDLINE];
-    if (sp != NULL)
-    {
-        snprintf(newcmd, sizeof(newcmd),
-                 "%.*s %s %s",
-                 (int)(sp - rhs),
-                 rhs, lhs, sp + 1);
-    }
-    else
-    {
-        snprintf(newcmd, sizeof(newcmd),
-                 "%s %s", rhs, lhs);
-    }
-    strncpy(data.CLIcmdline, newcmd,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    data.CLIcmdline[
-        STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-    *retval = CLI_execute_line();
-    data.CMDexecuted = 1;
-    return 1;
-}
-
-
-/**
- * @brief Split command at top-level pipe operator.
- *
- * Handles "cmd1 | cmd2" by capturing stdout of the
- * left command into a temp file and feeding it as
- * stdin to the right command. Only matches single
- * '|', not '||' or '|>'.
- *
- * Falls through (returns 0) if the right side is
- * not a registered milk command — the later popen-
- * based handler deals with shell pipes.
- *
- * @param[out] retval  Set to the return value if
- *                     the line was consumed
- * @return 1 if the line was consumed, 0 otherwise
- */
-static int cli_split_pipe(errno_t *retval)
-{
-    const char *src = data.CLIcmdline;
-    int depth = 0;
-    int in_sq = 0;
-    int in_dq = 0;
-    int pipe_pos = -1;
-
-    for (int si = 0; src[si] != '\0'; si++)
-    {
-        char c = src[si];
-        if (c == '\'' && !in_dq)
-        {
-            in_sq = !in_sq;
-        }
-        else if (c == '"' && !in_sq)
-        {
-            in_dq = !in_dq;
-        }
-        else if (!in_sq && !in_dq)
-        {
-            if (c == '(')
-            {
-                depth++;
-            }
-            else if (c == ')' && depth > 0)
-            {
-                depth--;
-            }
-            else if (depth == 0
-                     && c == '|'
-                     && src[si + 1] != '|'
-                     && src[si + 1] != '>')
-            {
-                pipe_pos = si;
-                break;
-            }
-        }
-    }
-    if (pipe_pos < 0)
-    {
-        return 0;
-    }
-
-    char left[STRINGMAXLEN_CLICMDLINE];
-    strncpy(left, data.CLIcmdline,
-            (size_t) pipe_pos);
-    left[pipe_pos] = '\0';
-    const char *rp =
-        data.CLIcmdline + pipe_pos + 1;
-    while (*rp == ' ' || *rp == '\t')
-    {
-        rp++;
-    }
-    char right[STRINGMAXLEN_CLICMDLINE];
-    strncpy(right, rp,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    right[STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-    /* Check if right side is a milk command;
-     * if not, fall through for popen handler */
-    {
-        char rword[200];
-        int rw = 0;
-        const char *rr = right;
-        while (*rr == ' ' || *rr == '\t')
-        {
-            rr++;
-        }
-        while (*rr != '\0'
-               && *rr != ' '
-               && *rr != '\t'
-               && rw < (int) sizeof(rword) - 1)
-        {
-            rword[rw++] = *rr++;
-        }
-        rword[rw] = '\0';
-        if (!cli_is_command(rword))
-        {
-            return 0;
-        }
-    }
-
-    /* Capture left stdout into tmpfile */
-    FILE *tmpfp = tmpfile();
-    if (tmpfp != NULL)
-    {
-        fflush(stdout);
-        int saved_stdout = dup(STDOUT_FILENO);
-        dup2(fileno(tmpfp), STDOUT_FILENO);
-
-        strncpy(data.CLIcmdline, left,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        CLI_execute_line();
-
-        fflush(stdout);
-        dup2(saved_stdout, STDOUT_FILENO);
-        close(saved_stdout);
-
-        /* Feed tmpfile as stdin to right */
-        rewind(tmpfp);
-        int saved_stdin = dup(STDIN_FILENO);
-        dup2(fileno(tmpfp), STDIN_FILENO);
-
-        strncpy(data.CLIcmdline, right,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        *retval = CLI_execute_line();
-
-        dup2(saved_stdin, STDIN_FILENO);
-        close(saved_stdin);
-        fclose(tmpfp);
-        return 1;
-    }
-    return 0;
-}
-
-
-/**
- * @brief Execute an external command with minimal overhead.
- *
- * Prefers a direct fork+exec via posix_spawnp() for simple
- * commands (no shell metacharacters), saving the extra
- * /bin/sh layer that system() would otherwise spawn.
- * Falls back to "/bin/sh -c cmd" only when metacharacters
- * are detected (pipes, redirects, glob, etc.).
- *
- * @param cmd  Fully expanded command string
- * @return  Exit status (0 = success), 127 if not found,
- *          -1 on spawn error
- */
-static int cli_run_external(
-    const char *cmd
-)
-{
-    extern char **environ;
-
-    /* Detect characters that require /bin/sh parsing */
-    static const char sh_meta[] =
-        "|&;<>()`\\\"'\n*?[{$";
-    int needs_shell = 0;
-    for(const char *cp = cmd; *cp; cp++)
-    {
-        if(strchr(sh_meta, *cp))
-        {
-            needs_shell = 1;
-            break;
-        }
-    }
-
-    pid_t pid  = -1;
-    int   ret  = -1;
-    int   status = 0;
-
-    if(needs_shell)
-    {
-        /* Shell metacharacters present:
-         * spawn /bin/sh -c cmd — equivalent to
-         * system() but using posix_spawn for
-         * cleaner signal handling */
-        char *const sh_args[] = {
-            "/bin/sh", "-c",
-            (char *) cmd, NULL
-        };
-        ret = posix_spawn(
-            &pid, "/bin/sh",
-            NULL, NULL,
-            sh_args, environ);
-    }
-    else
-    {
-        /* Simple command: tokenize on whitespace
-         * and exec directly, skipping the shell
-         * layer entirely */
-        char buf[STRINGMAXLEN_CLICMDLINE];
-        strncpy(buf, cmd, sizeof(buf) - 1);
-        buf[sizeof(buf) - 1] = '\0';
-
-        char *argv[256];
-        int   argc = 0;
-        char *tok  = strtok(buf, " \t");
-        while(tok != NULL && argc < 255)
-        {
-            argv[argc++] = tok;
-            tok = strtok(NULL, " \t");
-        }
-        argv[argc] = NULL;
-
-        if(argc == 0)
-        {
-            return 0;
-        }
-
-        ret = posix_spawnp(
-            &pid, argv[0],
-            NULL, NULL,
-            argv, environ);
-        if(ret != 0)
-        {
-            if(ret == ENOENT)
-            {
-                fprintf(stderr,
-                        "milk-cli: %s: not found\n",
-                        argv[0]);
-                return 127; /* command not found */
-            }
-            else if(ret == EACCES)
-            {
-                fprintf(stderr,
-                        "milk-cli: %s: permission"
-                        " denied\n",
-                        argv[0]);
-                return 126; /* permission denied */
-            }
-            else
-            {
-                fprintf(stderr,
-                        "milk-cli: %s: %s\n",
-                        argv[0], strerror(ret));
-                return 1;
-            }
-        }
-    }
-
-    if(ret != 0)
-    {
-        return -1;
-    }
-
-    if(waitpid(pid, &status, 0) == -1)
-    {
-        return -1;
-    }
-
-    if(WIFEXITED(status))
-    {
-        return WEXITSTATUS(status);
-    }
-    if(WIFSIGNALED(status))
-    {
-        return 128 + WTERMSIG(status);
-    }
-    return -1;
-}
-
-
-/**
- * @brief Handle subshell execution: (cmd1; cmd2)
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_subshell(
-    errno_t *retval
-)
-{
-    const char *sp = data.CLIcmdline;
-    int spl = (int) strlen(sp);
-    if (spl < 3
-        || sp[0] != '('
-        || sp[spl - 1] != ')')
-    {
-        return 0;
-    }
-    char sbuf[STRINGMAXLEN_CLICMDLINE];
-    memcpy(sbuf, sp + 1,
-           (size_t)(spl - 2));
-    sbuf[spl - 2] = '\0';
-    pid_t spid = fork();
-    if (spid == 0)
-    {
-        char *tok = strtok(sbuf, ";");
-        while (tok != NULL)
-        {
-            const char *st = tok;
-            while (*st == ' '
-                   || *st == '\t')
-            {
-                st++;
-            }
-            if (*st != '\0')
-            {
-                CLI_execute_string(
-                    (char *) st);
-            }
-            tok = strtok(NULL, ";");
-        }
-        _exit(0);
-    }
-    else if (spid > 0)
-    {
-        int wst;
-        waitpid(spid, &wst, 0);
-        cli_last_retval =
-            WEXITSTATUS(wst);
-    }
-    *retval = 0;
-    return 1;
-}
-
-
-/**
- * @brief Handle late here-string (pipe-based).
- *
- * Alternative here-string handler that uses pipe()
- * instead of tmpfile() for feeding text to stdin.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_herestring_late(
-    errno_t *retval
-)
-{
-    const char *hs = strstr(
-        data.CLIcmdline, "<<<");
-    if (hs == NULL)
-    {
-        return 0;
-    }
-
-    char hcmd[STRINGMAXLEN_CLICMDLINE];
-    int hcl = (int)(hs - data.CLIcmdline);
-    if (hcl >= STRINGMAXLEN_CLICMDLINE)
-    {
-        hcl = STRINGMAXLEN_CLICMDLINE - 1;
-    }
-    memcpy(hcmd, data.CLIcmdline,
-           (size_t) hcl);
-    hcmd[hcl] = '\0';
-    while (hcl > 0
-           && (hcmd[hcl - 1] == ' '
-               || hcmd[hcl - 1] == '\t'))
-    {
-        hcmd[--hcl] = '\0';
-    }
-    const char *tp = hs + 3;
-    while (*tp == ' ' || *tp == '\t')
-    {
-        tp++;
-    }
-    char htxt[1024];
-    strncpy(htxt, tp, sizeof(htxt) - 1);
-    htxt[sizeof(htxt) - 1] = '\0';
-    int htl = (int) strlen(htxt);
-    if (htl >= 2
-        && ((htxt[0] == '"'
-             && htxt[htl - 1] == '"')
-            || (htxt[0] == '\''
-                && htxt[htl - 1] == '\'')))
-    {
-        htxt[htl - 1] = '\0';
-        memmove(htxt, htxt + 1,
-                (size_t)(htl - 1));
-    }
-    int pfd[2];
-    if (pipe(pfd) == 0)
-    {
-        ssize_t wr_ignore;
-        wr_ignore = write(pfd[1], htxt,
-                          strlen(htxt));
-        wr_ignore = write(pfd[1], "\n", 1);
-        (void) wr_ignore;
-        close(pfd[1]);
-        int sv = dup(STDIN_FILENO);
-        dup2(pfd[0], STDIN_FILENO);
-        close(pfd[0]);
-        CLI_execute_string(hcmd);
-        dup2(sv, STDIN_FILENO);
-        close(sv);
-    }
-    *retval = 0;
-    return 1;
-}
-
-
-/**
- * @brief Handle stderr redirection (2>&1, 2>file).
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_stderr_redir(
-    errno_t *retval
-)
-{
-    const char *se = strstr(
-        data.CLIcmdline, "2>");
-    if (se == NULL)
-    {
-        return 0;
-    }
-
-    char scmd[STRINGMAXLEN_CLICMDLINE];
-    int scl = (int)(se - data.CLIcmdline);
-    if (scl >= STRINGMAXLEN_CLICMDLINE)
-    {
-        scl = STRINGMAXLEN_CLICMDLINE - 1;
-    }
-    memcpy(scmd, data.CLIcmdline,
-           (size_t) scl);
-    scmd[scl] = '\0';
-    while (scl > 0
-           && (scmd[scl - 1] == ' '
-               || scmd[scl - 1] == '\t'))
-    {
-        scmd[--scl] = '\0';
-    }
-    const char *target = se + 2;
-    while (*target == ' '
-           || *target == '\t')
-    {
-        target++;
-    }
-    int sv_err = dup(STDERR_FILENO);
-    if (strncmp(target, "&1", 2) == 0)
-    {
-        dup2(STDOUT_FILENO, STDERR_FILENO);
-    }
-    else
-    {
-        char fname[256];
-        int fi = 0;
-        while (target[fi] != '\0'
-               && target[fi] != ' '
-               && target[fi] != '\t'
-               && fi < 254)
-        {
-            fname[fi] = target[fi];
-            fi++;
-        }
-        fname[fi] = '\0';
-        FILE *ef = fopen(fname, "w");
-        if (ef != NULL)
-        {
-            dup2(fileno(ef),
-                 STDERR_FILENO);
-            fclose(ef);
-        }
-    }
-    CLI_execute_string(scmd);
-    dup2(sv_err, STDERR_FILENO);
-    close(sv_err);
-    *retval = 0;
-    return 1;
-}
-
-
-/**
- * @brief Handle input redirection (cmd < file).
- *
- * Scans for unquoted < (not << or <<<), redirects
- * stdin from the named file, and re-executes.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_input_redir(
-    errno_t *retval
-)
-{
-    const char *cl4 = data.CLIcmdline;
-    int in_sq4 = 0, in_dq4 = 0, depth4 = 0;
-    int inr_pos = -1;
-
-    for (int ri = 0; cl4[ri] != '\0'; ri++)
-    {
-        if (cl4[ri] == '\'' && !in_dq4)
-        {
-            in_sq4 = !in_sq4;
-        }
-        else if (cl4[ri] == '"' && !in_sq4)
-        {
-            in_dq4 = !in_dq4;
-        }
-        else if (!in_sq4 && !in_dq4)
-        {
-            if (cl4[ri] == '(')
-            {
-                depth4++;
-            }
-            else if (cl4[ri] == ')'
-                     && depth4 > 0)
-            {
-                depth4--;
-            }
-            else if (depth4 == 0
-                     && cl4[ri] == '<'
-                     && cl4[ri + 1] != '<')
-            {
-                inr_pos = ri;
-                break;
-            }
-        }
-    }
-    if (inr_pos < 0)
-    {
-        return 0;
-    }
-
-    int fst = inr_pos + 1;
-    while (data.CLIcmdline[fst] == ' '
-           || data.CLIcmdline[fst] == '\t')
-    {
-        fst++;
-    }
-    char infile[512];
-    int ifi = 0;
-    while (data.CLIcmdline[fst] != '\0'
-           && data.CLIcmdline[fst] != ' '
-           && data.CLIcmdline[fst] != '\t'
-           && ifi < 511)
-    {
-        infile[ifi++] =
-            data.CLIcmdline[fst++];
-    }
-    infile[ifi] = '\0';
-
-    data.CLIcmdline[inr_pos] = '\0';
-    {
-        int cl5 = inr_pos - 1;
-        while (cl5 >= 0
-               && (data.CLIcmdline[cl5] == ' '
-                   || data.CLIcmdline[cl5]
-                   == '\t'))
-        {
-            data.CLIcmdline[cl5--] = '\0';
-        }
-    }
-
-    FILE *ifp = NULL;
-    int is_stream = 0;
-    char tempname[256] = "";
-    
-    if (strncmp(infile, "@S:", 3) == 0) {
-        is_stream = 1;
-        char *sname = infile + 3;
-        IMAGE *img = (IMAGE*) malloc(sizeof(IMAGE));
-        if (ImageStreamIO_read_sharedmem_image_toIMAGE(sname, img) == 0) {
-            snprintf(tempname, sizeof(tempname), "/tmp/milk_cli_inredir_XXXXXX");
-            int fd = mkstemp(tempname);
-            if (fd >= 0) {
-                FILE *tf = fdopen(fd, "w");
-                if (tf) {
-                    size_t bytes = ImageStreamIO_typesize(img->md->datatype) * img->md->nelement;
-                    fwrite(img->array.raw, 1, bytes, tf);
-                    fclose(tf);
-                } else {
-                    close(fd);
-                }
-                ifp = fopen(tempname, "r");
-                unlink(tempname);
-            }
-            ImageStreamIO_closeIm(img);
-        } else {
-            printf("stream redirection: stream %s not found\n", sname);
-        }
-        free(img);
-    } else {
-        ifp = fopen(infile, "r");
-    }
-
-    if (ifp != NULL)
-    {
-        int sv_in = dup(STDIN_FILENO);
-        dup2(fileno(ifp), STDIN_FILENO);
-
-        *retval = CLI_execute_line();
-
-        dup2(sv_in, STDIN_FILENO);
-        close(sv_in);
-        fclose(ifp);
-        
-        return 1;
-    } else {
-        if (is_stream && tempname[0] != '\0') {
-            unlink(tempname);
-        }
-        // Redirection was detected and the command line has been modified.
-        // Treat this as consumed and signal failure to the caller.
-        *retval = RETURN_FAILURE;
-        return 1;
-    }
-}
-
-
-/**
- * @brief Split at semicolon or chaining ops.
- *
- * Scans for the first unquoted ; or && or ||,
- * splits the line, executes the first part,
- * then conditionally executes the rest.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_split_semicolon(
-    errno_t *retval
-)
-{
-    char fullline[STRINGMAXLEN_CLICMDLINE];
-    strncpy(fullline, data.CLIcmdline,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    fullline[STRINGMAXLEN_CLICMDLINE - 1] =
-        '\0';
-
-    int chain_type = 0; /* 1=; 2=&& 3=|| */
-    int chain_off = -1;
-    int chain_len = 0;
-    for (int ci = 0;
-         fullline[ci] != '\0'; ci++)
-    {
-        if (fullline[ci] == '"')
-        {
-            ci++;
-            while (fullline[ci] != '\0'
-                   && fullline[ci] != '"')
-            {
-                ci++;
-            }
-            if (fullline[ci] == '\0')
-            {
-                break;
-            }
-            continue;
-        }
-        if (fullline[ci] == ';')
-        {
-            chain_type = 1;
-            chain_off = ci;
-            chain_len = 1;
-            break;
-        }
-        if (fullline[ci] == '&'
-            && fullline[ci + 1] == '&')
-        {
-            chain_type = 2;
-            chain_off = ci;
-            chain_len = 2;
-            break;
-        }
-        if (fullline[ci] == '|'
-            && fullline[ci + 1] == '|')
-        {
-            chain_type = 3;
-            chain_off = ci;
-            chain_len = 2;
-            break;
-        }
-    }
-    if (chain_off < 0)
-    {
-        return 0;
-    }
-
-    fullline[chain_off] = '\0';
-    strncpy(data.CLIcmdline, fullline,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    data.CLIcmdline[
-        STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-    errno_t ret1 = CLI_execute_line();
-
-    int run_rest = 0;
-    if (chain_type == 1)
-    {
-        run_rest = 1;
-    }
-    else if (chain_type == 2)
-    {
-        run_rest =
-            (ret1 == RETURN_SUCCESS) ? 1 : 0;
-    }
-    else if (chain_type == 3)
-    {
-        run_rest =
-            (ret1 != RETURN_SUCCESS) ? 1 : 0;
-    }
-    if (run_rest)
-    {
-        const char *rest =
-            fullline + chain_off + chain_len;
-        while (*rest == ' '
-               || *rest == '\t')
-        {
-            rest++;
-        }
-        if (*rest != '\0')
-        {
-            strncpy(data.CLIcmdline, rest,
-                    STRINGMAXLEN_CLICMDLINE
-                    - 1);
-            data.CLIcmdline[
-                STRINGMAXLEN_CLICMDLINE - 1]
-                = '\0';
-            CLI_execute_line();
-        }
-    }
-    *retval = RETURN_SUCCESS;
-    return 1;
-}
-
-
-/**
- * @brief Handle output redirection (> and >>).
- *
- * Scans for unquoted > or >> outside parens/quotes,
- * extracts the filename, redirects stdout, executes
- * the truncated command, then restores stdout.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_output_redir(
-    errno_t *retval
-)
-{
-    int redir_mode = 0; /* 1=trunc 2=append */
-    int redir_pos = -1;
-    int in_sq2 = 0, in_dq2 = 0, depth2 = 0;
-    const char *cl2 = data.CLIcmdline;
-
-    for (int ri = 0; cl2[ri] != '\0'; ri++)
-    {
-        if (cl2[ri] == '\'' && !in_dq2)
-        {
-            in_sq2 = !in_sq2;
-        }
-        else if (cl2[ri] == '"' && !in_sq2)
-        {
-            in_dq2 = !in_dq2;
-        }
-        else if (!in_sq2 && !in_dq2)
-        {
-            if (cl2[ri] == '(')
-            {
-                depth2++;
-            }
-            else if (cl2[ri] == ')'
-                     && depth2 > 0)
-            {
-                depth2--;
-            }
-            else if (depth2 == 0
-                     && cl2[ri] == '>')
-            {
-                if (cl2[ri + 1] == '>')
-                {
-                    redir_mode = 2;
-                }
-                else
-                {
-                    redir_mode = 1;
-                }
-                redir_pos = ri;
-            }
-        }
-    }
-    if (redir_pos < 0)
-    {
-        return 0;
-    }
-
-    int fstart = redir_pos
-                 + ((redir_mode == 2) ? 2 : 1);
-    while (data.CLIcmdline[fstart] == ' '
-           || data.CLIcmdline[fstart] == '\t')
-    {
-        fstart++;
-    }
-    char rfile[512];
-    int fi = 0;
-    while (data.CLIcmdline[fstart] != '\0'
-           && data.CLIcmdline[fstart] != ' '
-           && data.CLIcmdline[fstart] != '\t'
-           && fi < 511)
-    {
-        rfile[fi++] =
-            data.CLIcmdline[fstart++];
-    }
-    rfile[fi] = '\0';
-
-    /* Truncate cmd at redir */
-    data.CLIcmdline[redir_pos] = '\0';
-    {
-        int cl3 = redir_pos - 1;
-        while (cl3 >= 0
-               && (data.CLIcmdline[cl3] == ' '
-                   || data.CLIcmdline[cl3]
-                   == '\t'))
-        {
-            data.CLIcmdline[cl3--] = '\0';
-        }
-    }
-
-    FILE *rfp = NULL;
-    int is_fps = 0;
-    int is_stream = 0;
-    char tempname[256];
-    
-    if (strncmp(rfile, "@F:", 3) == 0) {
-        is_fps = 1;
-        snprintf(tempname, sizeof(tempname), "/tmp/milk_cli_fredir_XXXXXX");
-        int fd = mkstemp(tempname);
-        if (fd >= 0) {
-            rfp = fdopen(fd, (redir_mode == 2) ? "a" : "w");
-            if (!rfp) close(fd);
-        }
-    } else if (strncmp(rfile, "@S:", 3) == 0) {
-        is_stream = 1;
-        snprintf(tempname, sizeof(tempname), "/tmp/milk_cli_sredir_XXXXXX");
-        int fd = mkstemp(tempname);
-        if (fd >= 0) {
-            rfp = fdopen(fd, (redir_mode == 2) ? "a" : "w");
-            if (!rfp) close(fd);
-        }
-    } else {
-        rfp = fopen(rfile, (redir_mode == 2) ? "a" : "w");
-    }
-
-    if (rfp != NULL)
-    {
-        fflush(stdout);
-        int sv_out = dup(STDOUT_FILENO);
-        dup2(fileno(rfp), STDOUT_FILENO);
-
-        *retval = CLI_execute_line();
-
-        fflush(stdout);
-        dup2(sv_out, STDOUT_FILENO);
-        close(sv_out);
-        fclose(rfp);
-        
-        if (is_fps) {
-            char *fpspath = rfile + 3;
-            char *dot = strchr(fpspath, '.');
-            if (dot != NULL) {
-                *dot = '\0';
-                char *fpsname = fpspath;
-                char *param = dot + 1;
-                
-                FILE *tf = fopen(tempname, "r");
-                if (tf) {
-                    char valbuf[2048] = {0};
-                    size_t rn = fread(valbuf, 1, sizeof(valbuf)-1, tf);
-                    valbuf[rn] = '\0';
-                    fclose(tf);
-                    
-                    while(rn > 0 && (valbuf[rn-1] == '\n' || valbuf[rn-1] == '\r')) {
-                        valbuf[--rn] = '\0';
-                    }
-                    
-                    FUNCTION_PARAMETER_STRUCT fps_struct;
-                    if (function_parameter_struct_connect(fpsname, &fps_struct, FPSCONNECT_SIMPLE) != -1 && fps_struct.parray != NULL) {
-                        int pidx = functionparameter_GetParamIndex(&fps_struct, param);
-                        if (pidx < 0) {
-                            char dotname[512];
-                            snprintf(dotname, sizeof(dotname), ".%s", param);
-                            pidx = functionparameter_GetParamIndex(&fps_struct, dotname);
-                        }
-                        if (pidx >= 0) {
-                            const char *kw = fps_struct.parray[pidx].keyword[0];
-                            switch (fps_struct.parray[pidx].type) {
-                                case FPTYPE_INT32:
-                                    functionparameter_SetParamValue_INT32(&fps_struct, kw, strtol(valbuf, NULL, 10)); break;
-                                case FPTYPE_UINT32:
-                                    functionparameter_SetParamValue_UINT32(&fps_struct, kw, strtoul(valbuf, NULL, 10)); break;
-                                case FPTYPE_INT64:
-                                case FPTYPE_PID:
-                                    functionparameter_SetParamValue_INT64(&fps_struct, kw, strtoll(valbuf, NULL, 10)); break;
-                                case FPTYPE_UINT64:
-                                    functionparameter_SetParamValue_UINT64(&fps_struct, kw, strtoull(valbuf, NULL, 10)); break;
-                                case FPTYPE_FLOAT32:
-                                    functionparameter_SetParamValue_FLOAT32(&fps_struct, kw, strtof(valbuf, NULL)); break;
-                                case FPTYPE_FLOAT64:
-                                    functionparameter_SetParamValue_FLOAT64(&fps_struct, kw, strtod(valbuf, NULL)); break;
-                                case FPTYPE_TIMESPEC:
-                                    functionparameter_SetParamValue_TIMESPEC(&fps_struct, kw, strtof(valbuf, NULL)); break;
-                                case FPTYPE_ONOFF:
-                                    if (strcasecmp(valbuf, "ON") == 0 || strcmp(valbuf, "1") == 0) {
-                                        functionparameter_SetParamValue_ONOFF(&fps_struct, kw, 1);
-                                    } else if (strcasecmp(valbuf, "OFF") == 0 || strcmp(valbuf, "0") == 0) {
-                                        functionparameter_SetParamValue_ONOFF(&fps_struct, kw, 0);
-                                    }
-                                    break;
-                                default:
-                                    functionparameter_SetParamValue_STRING(&fps_struct, kw, valbuf); break;
-                            }
-                        } else {
-                            printf("fps redirection: param %s not found in %s\n", param, fpsname);
-                        }
-                        function_parameter_struct_disconnect(&fps_struct);
-                    } else {
-                        printf("fps redirection: could not connect to %s\n", fpsname);
-                    }
-                }
-            } else {
-                printf("fps redirection: format must be @F:fpsname.param\n");
-            }
-            unlink(tempname);
-        } else if (is_stream) {
-            char *sname = rfile + 3;
-            IMAGE *img = (IMAGE*) malloc(sizeof(IMAGE));
-            if (ImageStreamIO_read_sharedmem_image_toIMAGE(sname, img) == 0) {
-                FILE *tf = fopen(tempname, "r");
-                if (tf) {
-                    size_t bytes = ImageStreamIO_typesize(img->md->datatype) * img->md->nelement;
-                    size_t bread = fread(img->array.raw, 1, bytes, tf);
-                    (void)bread;
-                    fclose(tf);
-                }
-                ImageStreamIO_UpdateIm(img);
-
-                ImageStreamIO_closeIm(img);
-            } else {
-                printf("stream redirection: stream %s not found\n", sname);
-            }
-            free(img);
-            unlink(tempname);
-        }
-        return 1;
-    } else {
-        if ((is_fps || is_stream) && tempname[0] != '\0') {
-            unlink(tempname);
-        }
-    }
-    return 0;
-}
-
-
-/**
- * @brief Handle here-string syntax (<<<).
- *
- * Parses "cmd <<< text", strips quotes from text,
- * writes it to a tmpfile, redirects stdin, and
- * executes the command.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_herestring_early(
-    errno_t *retval
-)
-{
-    char *hs = strstr(
-        data.CLIcmdline, "<<<");
-    if (hs == NULL)
-    {
-        return 0;
-    }
-
-    *hs = '\0';
-    const char *hsval = hs + 3;
-    while (*hsval == ' ' || *hsval == '\t')
-    {
-        hsval++;
-    }
-    int hvlen = (int) strlen(hsval);
-    char hvbuf[STRINGMAXLEN_CLICMDLINE];
-    if (hvlen >= 2
-        && ((hsval[0] == '"'
-             && hsval[hvlen - 1] == '"')
-            || (hsval[0] == '\''
-                && hsval[hvlen - 1] == '\'')))
-    {
-        memcpy(hvbuf, hsval + 1,
-               (size_t)(hvlen - 2));
-        hvbuf[hvlen - 2] = '\0';
-    }
-    else
-    {
-        strncpy(hvbuf, hsval,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        hvbuf[STRINGMAXLEN_CLICMDLINE - 1] =
-            '\0';
-    }
-
-    FILE *hsfp = tmpfile();
-    if (hsfp != NULL)
-    {
-        fprintf(hsfp, "%s\n", hvbuf);
-        rewind(hsfp);
-        int sv_in = dup(STDIN_FILENO);
-        dup2(fileno(hsfp), STDIN_FILENO);
-
-        *retval = CLI_execute_line();
-
-        dup2(sv_in, STDIN_FILENO);
-        close(sv_in);
-        fclose(hsfp);
-        return 1;
-    }
-    return 0;
-}
-
-
-/**
- * @brief Handle background execution (cmd &).
- *
- * If the command ends with an unquoted trailing &
- * (not &&), forks a child process to execute the
- * command and returns immediately.
- *
- * @param[out] retval  Return value if consumed
- * @return 1 if consumed, 0 otherwise
- */
-static int cli_handle_background(
-    errno_t *retval
-)
-{
-    int ll = (int) strlen(data.CLIcmdline);
-    int bi = ll - 1;
-    while (bi >= 0
-           && (data.CLIcmdline[bi] == ' '
-               || data.CLIcmdline[bi] == '\t'))
-    {
-        bi--;
-    }
-    if (bi < 0
-        || data.CLIcmdline[bi] != '&'
-        || (bi > 0
-            && data.CLIcmdline[bi - 1] == '&'))
-    {
-        return 0;
-    }
-
-    data.CLIcmdline[bi] = '\0';
-    pid_t cpid = fork();
-    if (cpid == 0)
-    {
-        CLI_execute_string(data.CLIcmdline);
-        _exit(0);
-    }
-    else if (cpid > 0)
-    {
-        printf("[bg] %d\n", (int) cpid);
-        char pb[32];
-        snprintf(pb, sizeof(pb),
-                 "%d", (int) cpid);
-        cli_var_set("!", pb);
-    }
-    *retval = 0;
-    return 1;
-}
-
-
-/**
- * @brief Rewrite dot-source syntax to source command.
- *
- * Transforms ". file" into "source file" in
- * data.CLIcmdline so the normal source handler
- * processes it.
- *
- * Always returns 0 (not consumed) — the rewritten
- * line is passed through to subsequent stages.
- */
-static int cli_rewrite_dot_source(void)
-{
-    const char *p = data.CLIcmdline;
-    while (*p == ' ' || *p == '\t')
-    {
-        p++;
-    }
-    if (p[0] == '.' && p[1] == ' ')
-    {
-        char tmp[STRINGMAXLEN_CLICMDLINE];
-        snprintf(tmp,
-                 STRINGMAXLEN_CLICMDLINE,
-                 "source %s", p + 2);
-        strncpy(data.CLIcmdline, tmp,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        data.CLIcmdline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-    }
-    return 0;
-}
-
 
 /**
  * @brief Central command dispatcher — the main entry
